+++ src/main/java/org/dspace/pack/bagit/CollectionPacker.java
@@ -45,7 +45,7 @@ import org.dspace.pack.PackerFactory;
 import org.dspace.pack.bagit.xml.metadata.Metadata;
 import org.dspace.pack.bagit.xml.metadata.Value;
 import org.dspace.pack.bagit.xml.policy.Policies;
-import org.dspace.pack.bagit.xml.site.DSpaceRoles;
+import org.dspace.pack.bagit.xml.roles.DSpaceRoles;
 
 /**
  * CollectionPacker packs and unpacks Collection AIPs in BagIt bags
