+++ src/main/java/org/dspace/pack/bagit/xml/roles/AssociatedGroup.java
@@ -1,4 +1,4 @@
-package org.dspace.pack.bagit.xml.site;
+package org.dspace.pack.bagit.xml.roles;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
