+++ src/main/java/org/dspace/pack/bagit/xml/roles/DSpaceRoles.java
@@ -1,4 +1,4 @@
-package org.dspace.pack.bagit.xml.site;
+package org.dspace.pack.bagit.xml.roles;
 
 import java.util.HashSet;
 import java.util.Set;
