+++ src/test/java/org/dspace/pack/bagit/BagItRolesUtilTest.java
@@ -35,10 +35,10 @@ import org.dspace.eperson.Group;
 import org.dspace.eperson.factory.EPersonServiceFactory;
 import org.dspace.eperson.service.EPersonService;
 import org.dspace.eperson.service.GroupService;
-import org.dspace.pack.bagit.xml.site.AssociatedGroup;
-import org.dspace.pack.bagit.xml.site.DSpaceRoles;
-import org.dspace.pack.bagit.xml.site.Member;
-import org.dspace.pack.bagit.xml.site.Person;
+import org.dspace.pack.bagit.xml.roles.AssociatedGroup;
+import org.dspace.pack.bagit.xml.roles.DSpaceRoles;
+import org.dspace.pack.bagit.xml.roles.Member;
+import org.dspace.pack.bagit.xml.roles.Person;
 import org.junit.Before;
 import org.junit.Test;
