+++ src/test/java/org/dspace/pack/bagit/xml/roles/DSpaceRolesTest.java
@@ -1,4 +1,4 @@
-package org.dspace.pack.bagit.xml.site;
+package org.dspace.pack.bagit.xml.roles;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
